Moved manifold laplacian functions and utils file in utils folder
